WIP - fix bad responses
