+++ app/routes/chatbots.$chatbotId.chat.$chatId/chat-input.tsx
