+++ app/routes/chatbots.$chatbotId.chat.$chatId/route.tsx
@@ -1,6 +1,9 @@
 import { LoaderFunctionArgs, json } from "@remix-run/node";
-import Chat from "~/components/chat/chat";
+import { ShouldRevalidateFunctionArgs } from "@remix-run/react";
 import { prisma } from "~/db.server";
+import Chat from "./chat";
+
+// when the data connector modal renders, we submit a reset, which triggers this to revalidate
 
 export const loader = async ({ request, params }: LoaderFunctionArgs) => {
   const { chatId, chatbotId } = params;
@@ -14,7 +17,8 @@ export const loader = async ({ request, params }: LoaderFunctionArgs) => {
       ? process.env.DEV_BASE
       : process.env.PROD_BASE;
 
-  const [messages, chatbotRes] = await Promise.all([
+  // TODO - can probably put this in a prisma transaction
+  const [messages, chatbotRes, chat] = await Promise.all([
     prisma.message.findMany({
       where: { chatId },
       orderBy: { createdAt: "asc" },
@@ -28,6 +32,36 @@ export const loader = async ({ request, params }: LoaderFunctionArgs) => {
       },
     }),
     fetch(`${BASE_URL}/api/chatbot/${chatbotId}`),
+    prisma.chat.findUnique({
+      where: { id: chatId },
+      include: {
+        messages: {
+          include: {
+            form: {
+              include: {
+                elements: true,
+              },
+            },
+            formSubmission: true,
+          },
+          orderBy: {
+            createdAt: "asc",
+          },
+        },
+        _count: {
+          select: {
+            messages: {
+              where: {
+                seenByUser: false,
+                role: {
+                  not: "user",
+                },
+              },
+            },
+          },
+        },
+      },
+    }),
   ]);
 
   if (!chatbotRes.ok) {
@@ -36,14 +70,28 @@ export const loader = async ({ request, params }: LoaderFunctionArgs) => {
 
   const { chatbot } = await chatbotRes.json();
 
+  console.log("revalidation");
+
   return json({
     chatbot,
-    messages,
+    loaderMessages: messages,
     BASE_URL,
     userMessage: new URL(request.url).searchParams.get("userMessage"),
+    loaderChat: chat,
   });
 };
 
+export function shouldRevalidate({
+  formData,
+  defaultShouldRevalidate,
+}: ShouldRevalidateFunctionArgs) {
+  if (formData?.get("intent") === "reset") {
+    return false;
+  }
+
+  return defaultShouldRevalidate;
+}
+
 export default function ChatRoute() {
-  return <Chat key="chat" />;
+  return <Chat />;
 }
