+++ app/routes/chatbots.$chatbotId.data._index/ui/modal.tsx
@@ -1,7 +1,7 @@
 import { Button } from "~/components/ui/button";
 import { Dialog, DialogContent, DialogTrigger } from "~/components/ui/dialog";
 
-import { useSubmit } from "@remix-run/react";
+import { useParams, useSubmit } from "@remix-run/react";
 import { Plus } from "lucide-react";
 import { useEffect, useState } from "react";
 import { STEPS } from "~/utils/types";
@@ -13,13 +13,20 @@ import Website from "./website";
 
 export function DialogDemo({
   submit,
+  prefillStep,
+  prefillQuestion,
 }: {
   submit: ReturnType<typeof useSubmit>;
+  prefillStep?: string;
+  prefillQuestion?: string;
 }) {
   // we should pass these two down - as parentmodalcontrols ...
   // maybe we should pass the submit down as well?
-  const [step, setStep] = useState(STEPS.SELECT_TYPE);
+  const [step, setStep] = useState(prefillStep || STEPS.SELECT_TYPE);
   const [open, setOpen] = useState(false);
+  const { chatbotId } = useParams();
+
+  console.log("prefillQuestion", open, prefillQuestion, prefillStep);
 
   // on mount - stop all jobs that can come from the modal (website, file, blank. ...)
   // and on unmount?
@@ -29,6 +36,8 @@ export function DialogDemo({
       { intent: "reset" },
       {
         method: "post",
+        action: `/chatbots/${chatbotId}/data?index`,
+        navigate: false,
 
         // unstable_flushSync: true
       },
@@ -38,7 +47,7 @@ export function DialogDemo({
   return (
     <Dialog open={open} onOpenChange={setOpen}>
       <DialogTrigger asChild>
-        <Button onClick={() => setStep(STEPS.SELECT_TYPE)}>
+        <Button onClick={() => setOpen(true)}>
           <Plus className="mr-2 h-4 w-4" />{" "}
           <span className="text-md">Add data</span>
         </Button>
@@ -55,7 +64,12 @@ export function DialogDemo({
           <BlankUpload setStep={setStep} setOpen={setOpen} submit={submit} />
         ) : null}
         {step === STEPS.QA ? (
-          <QA setStep={setStep} setOpen={setOpen} submit={submit} />
+          <QA
+            setStep={setStep}
+            setOpen={setOpen}
+            submit={submit}
+            prefillQuestion={prefillQuestion}
+          />
         ) : null}
       </DialogContent>
     </Dialog>
