+++ package.json
@@ -95,6 +95,7 @@
     "axios": "^1.6.7",
     "bcryptjs": "^2.4.3",
     "bullmq": "^5.7.15",
+    "canvas-confetti": "^1.9.3",
     "chalk": "^5.3.0",
     "cheerio": "^1.0.0-rc.12",
     "chokidar": "^3.5.3",
@@ -238,6 +239,7 @@
     "@testing-library/jest-dom": "^6.1.4",
     "@types/async": "^3.2.24",
     "@types/bcryptjs": "^2.4.5",
+    "@types/canvas-confetti": "^1.6.4",
     "@types/compression": "^1.7.4",
     "@types/cors": "^2.8.17",
     "@types/eslint": "^8.44.6",
