+++ prisma/schema.prisma
@@ -268,6 +268,10 @@ model Message {
   flowId String?
   flow   Flow?   @relation(fields: [flowId], references: [id], onDelete: Cascade)
 
+  // WIP - postprocessing of messages for improving chatbot performance
+  didNotFulfillQuery Boolean? @default(false)
+  reasoning          String?
+
   @@index([chatId, role])
 }
